style(common.py): Moving wait_for from scylla_node
The method is generic and can be "call" from each file in the package
