+++ ccmlib/common.py
@@ -461,6 +461,37 @@ def scylla_extract_install_dir_and_mode(install_dir):
     return install_dir, scylla_mode
 
 
+def wait_for(func, timeout, first=0.0, step=1.0, text=None):
+    """
+    Wait until func() evaluates to True.
+
+    If func() evaluates to True before timeout expires, return the
+    value of func(). Otherwise return None.
+
+    :param func: Function that will be evaluated.
+    :param timeout: Timeout in seconds
+    :param first: Time to sleep before first attempt
+    :param step: Time to sleep between attempts in seconds
+    :param text: Text to print while waiting, for debug purposes
+    """
+    start_time = time.time()
+    end_time = time.time() + timeout
+
+    time.sleep(first)
+
+    while time.time() < end_time:
+        if text:
+            print_("%s (%f secs)" % (text, (time.time() - start_time)))
+
+        output = func()
+        if output:
+            return output
+
+        time.sleep(step)
+
+    return None
+
+
 def check_file_exists(path_to_file):
     if os.path.exists(path_to_file):
         return True
